requirements.txt
fastapi
uvicorn
openai
python-dotenv
pymongo
httpx
jsonlines
python-multipart
NRCLex
translate
mecab-python3
unidic-lite